Separacion
Separación de los distintos archivos en cabecera, main y fuente.
